+++ crates/networking/p2p/metrics.rs
@@ -1,8 +1,9 @@
+use core::fmt;
 use std::{
     collections::{BTreeMap, VecDeque},
     sync::{
         Arc, LazyLock,
-        atomic::{AtomicU64, Ordering},
+        atomic::{AtomicU8, AtomicU64, Ordering},
     },
     time::{Duration, SystemTime},
 };
@@ -11,7 +12,7 @@ use ethrex_common::H256;
 use prometheus::{Gauge, IntCounter, Registry};
 use tokio::sync::Mutex;
 
-use crate::rlpx::{error::RLPxError, p2p::DisconnectReason};
+use crate::rlpx::{error::PeerConnectionError, p2p::DisconnectReason};
 
 pub static METRICS: LazyLock<Metrics> = LazyLock::new(Metrics::default);
 
@@ -57,7 +58,7 @@ pub struct Metrics {
     // Common
     pub sync_head_block: AtomicU64,
     pub sync_head_hash: Arc<Mutex<H256>>,
-    pub current_step: Arc<Mutex<String>>,
+    pub current_step: Arc<CurrentStep>,
 
     // Headers
     pub headers_to_download: AtomicU64,
@@ -99,6 +100,89 @@ pub struct Metrics {
     pub bytecode_download_end_time: Arc<Mutex<Option<SystemTime>>>,
 
     start_time: SystemTime,
+}
+
+#[derive(Debug)]
+pub struct CurrentStep(AtomicU8);
+
+impl CurrentStep {
+    pub fn set(&self, value: CurrentStepValue) {
+        self.0.store(value.into(), Ordering::Relaxed);
+    }
+
+    pub fn get(&self) -> CurrentStepValue {
+        self.0.load(Ordering::Relaxed).into()
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+#[repr(u8)]
+pub enum CurrentStepValue {
+    None,
+    HealingStorage,
+    HealingState,
+    RequestingBytecodes,
+    RequestingAccountRanges,
+    RequestingStorageRanges,
+    DownloadingHeaders,
+    InsertingStorageRanges,
+    InsertingAccountRanges,
+    InsertingAccountRangesNoDb,
+}
+
+impl From<u8> for CurrentStepValue {
+    fn from(value: u8) -> Self {
+        match value {
+            1 => Self::HealingStorage,
+            2 => Self::HealingState,
+            3 => Self::RequestingBytecodes,
+            4 => Self::RequestingAccountRanges,
+            5 => Self::RequestingStorageRanges,
+            6 => Self::DownloadingHeaders,
+            7 => Self::InsertingStorageRanges,
+            8 => Self::InsertingAccountRanges,
+            9 => Self::InsertingAccountRangesNoDb,
+            _ => Self::None,
+        }
+    }
+}
+
+impl From<CurrentStepValue> for u8 {
+    fn from(value: CurrentStepValue) -> Self {
+        match value {
+            CurrentStepValue::None => 0,
+            CurrentStepValue::HealingStorage => 1,
+            CurrentStepValue::HealingState => 2,
+            CurrentStepValue::RequestingBytecodes => 3,
+            CurrentStepValue::RequestingAccountRanges => 4,
+            CurrentStepValue::RequestingStorageRanges => 5,
+            CurrentStepValue::DownloadingHeaders => 6,
+            CurrentStepValue::InsertingStorageRanges => 7,
+            CurrentStepValue::InsertingAccountRanges => 8,
+            CurrentStepValue::InsertingAccountRangesNoDb => 9,
+        }
+    }
+}
+
+impl fmt::Display for CurrentStepValue {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            CurrentStepValue::None => write!(f, "Unknown"),
+            CurrentStepValue::HealingStorage => write!(f, "Healing Storage"),
+            CurrentStepValue::HealingState => write!(f, "Healing State"),
+            CurrentStepValue::RequestingBytecodes => write!(f, "Requesting Bytecodes"),
+            CurrentStepValue::RequestingAccountRanges => write!(f, "Requesting Account Ranges"),
+            CurrentStepValue::RequestingStorageRanges => write!(f, "Requesting Storage Ranges"),
+            CurrentStepValue::DownloadingHeaders => write!(f, "Downloading Headers"),
+            CurrentStepValue::InsertingStorageRanges => {
+                write!(f, "Inserting Storage Ranges - \x1b[31mWriting to DB\x1b[0m")
+            }
+            CurrentStepValue::InsertingAccountRanges => {
+                write!(f, "Inserting Account Ranges - \x1b[31mWriting to DB\x1b[0m")
+            }
+            CurrentStepValue::InsertingAccountRangesNoDb => write!(f, "Inserting Account Ranges"),
+        }
+    }
 }
 
 impl Metrics {
@@ -195,7 +279,7 @@ impl Metrics {
             .and_modify(|count| *count -= 1);
     }
 
-    pub async fn record_new_rlpx_conn_failure(&self, reason: RLPxError) {
+    pub async fn record_new_rlpx_conn_failure(&self, reason: PeerConnectionError) {
         let mut failures_grouped_by_reason = self.connection_attempt_failures.lock().await;
 
         self.update_failures_grouped_by_reason(&mut failures_grouped_by_reason, &reason)
@@ -242,186 +326,204 @@ impl Metrics {
     pub async fn update_failures_grouped_by_reason(
         &self,
         failures_grouped_by_reason: &mut BTreeMap<String, u64>,
-        failure_reason: &RLPxError,
+        failure_reason: &PeerConnectionError,
     ) {
         match failure_reason {
-            RLPxError::HandshakeError(reason) => {
+            PeerConnectionError::HandshakeError(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("HandshakeError - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::StateError(reason) => {
+            PeerConnectionError::StateError(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("StateError - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::NoMatchingCapabilities() => {
+            PeerConnectionError::NoMatchingCapabilities() => {
                 failures_grouped_by_reason
                     .entry("NoMatchingCapabilities".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::Disconnected() => {
+            PeerConnectionError::Disconnected => {
                 failures_grouped_by_reason
                     .entry("Disconnected".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::DisconnectReceived(disconnect_reason) => {
+            PeerConnectionError::DisconnectReceived(disconnect_reason) => {
                 failures_grouped_by_reason
                     .entry(format!("DisconnectReceived - {disconnect_reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::DisconnectSent(disconnect_reason) => {
+            PeerConnectionError::DisconnectSent(disconnect_reason) => {
                 failures_grouped_by_reason
                     .entry(format!("DisconnectSent - {disconnect_reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::NotFound(reason) => {
+            PeerConnectionError::NotFound(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("NotFound - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::InvalidPeerId() => {
+            PeerConnectionError::InvalidPeerId() => {
                 failures_grouped_by_reason
                     .entry("InvalidPeerId".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::InvalidRecoveryId() => {
+            PeerConnectionError::InvalidRecoveryId() => {
                 failures_grouped_by_reason
                     .entry("InvalidRecoveryId".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::InvalidMessageLength() => {
+            PeerConnectionError::InvalidMessageLength() => {
                 failures_grouped_by_reason
                     .entry("InvalidMessageLength".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::MessageNotHandled(reason) => {
+            PeerConnectionError::ExpectedRequestId(reason) => {
+                failures_grouped_by_reason
+                    .entry(format!("ExpectedRequestId - {reason}"))
+                    .and_modify(|e| *e += 1)
+                    .or_insert(1);
+            }
+            PeerConnectionError::MessageNotHandled(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("MessageNotHandled - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::BadRequest(reason) => {
+            PeerConnectionError::BadRequest(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("BadRequest - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::RLPDecodeError(rlpdecode_error) => {
+            PeerConnectionError::RLPDecodeError(rlpdecode_error) => {
                 failures_grouped_by_reason
                     .entry(format!("RLPDecodeError - {rlpdecode_error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::RLPEncodeError(rlpencode_error) => {
+            PeerConnectionError::RLPEncodeError(rlpencode_error) => {
                 failures_grouped_by_reason
                     .entry(format!("RLPEncodeError - {rlpencode_error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::StoreError(store_error) => {
+            PeerConnectionError::StoreError(store_error) => {
                 failures_grouped_by_reason
                     .entry(format!("StoreError - {store_error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::CryptographyError(reason) => {
+            PeerConnectionError::CryptographyError(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("CryptographyError - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::BroadcastError(reason) => {
+            PeerConnectionError::BroadcastError(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("BroadcastError - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::RecvError(recv_error) => {
+            PeerConnectionError::RecvError(recv_error) => {
                 failures_grouped_by_reason
                     .entry(format!("RecvError - {recv_error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::SendMessage(reason) => {
+            PeerConnectionError::SendMessage(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("SendMessage - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::MempoolError(mempool_error) => {
+            PeerConnectionError::MempoolError(mempool_error) => {
                 failures_grouped_by_reason
                     .entry(format!("MempoolError - {mempool_error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::IoError(error) => {
+            PeerConnectionError::IoError(error) => {
                 failures_grouped_by_reason
                     .entry(format!("IoError - {error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::InvalidMessageFrame(reason) => {
+            PeerConnectionError::InvalidMessageFrame(reason) => {
                 failures_grouped_by_reason
                     .entry(format!("InvalidMessageFrame - {reason}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::IncompatibleProtocol => {
+            PeerConnectionError::IncompatibleProtocol => {
                 failures_grouped_by_reason
                     .entry("IncompatibleProtocol".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::InvalidBlockRange => {
+            PeerConnectionError::InvalidBlockRange => {
                 failures_grouped_by_reason
                     .entry("InvalidBlockRange".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::RollupStoreError(error) => {
+            PeerConnectionError::RollupStoreError(error) => {
                 failures_grouped_by_reason
                     .entry(format!("RollupStoreError - {error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::BlockchainError(error) => {
+            PeerConnectionError::BlockchainError(error) => {
                 failures_grouped_by_reason
                     .entry(format!("BlockchainError - {error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::InternalError(error) => {
+            PeerConnectionError::InternalError(error) => {
                 failures_grouped_by_reason
                     .entry(format!("InternalError - {error}"))
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::L2CapabilityNotNegotiated => {
+            PeerConnectionError::L2CapabilityNotNegotiated => {
                 failures_grouped_by_reason
                     .entry("L2CapabilityNotNegotiated".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::InvalidBlockRangeUpdate => {
+            PeerConnectionError::InvalidBlockRangeUpdate => {
                 failures_grouped_by_reason
                     .entry("InvalidBlockRangeUpdate".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
-            RLPxError::PeerTableError(error) => {
+            PeerConnectionError::PeerTableError(error) => {
                 failures_grouped_by_reason
                     .entry(format!("InternalError - {error}"))
+                    .and_modify(|e| *e += 1)
+                    .or_insert(1);
+            }
+            PeerConnectionError::Timeout => {
+                failures_grouped_by_reason
+                    .entry("Timeout".to_owned())
+                    .and_modify(|e| *e += 1)
+                    .or_insert(1);
+            }
+            PeerConnectionError::UnexpectedResponse(_, _) => {
+                failures_grouped_by_reason
+                    .entry("UnexpectedResponse".to_owned())
                     .and_modify(|e| *e += 1)
                     .or_insert(1);
             }
@@ -559,7 +661,7 @@ impl Default for Metrics {
             // Common
             sync_head_block: AtomicU64::new(0),
             sync_head_hash: Arc::new(Mutex::new(H256::default())),
-            current_step: Arc::new(Mutex::new("".to_string())),
+            current_step: Arc::new(CurrentStep(AtomicU8::new(0))),
 
             // Headers
             headers_to_download: AtomicU64::new(0),