+++ crates/networking/p2p/network.rs
@@ -1,11 +1,11 @@
 use crate::{
     discv4::{
-        peer_table::{PeerData, PeerTableHandle},
+        peer_table::{PeerData, PeerTable},
         server::{DiscoveryServer, DiscoveryServerError},
     },
     metrics::METRICS,
     rlpx::{
-        connection::server::{RLPxConnBroadcastSender, RLPxConnection},
+        connection::server::{PeerConnBroadcastSender, PeerConnection},
         initiator::RLPxInitiator,
         l2::l2_connection::P2PBasedContext,
         message::Message,
@@ -37,10 +37,10 @@ pub const MAX_MESSAGES_TO_BROADCAST: usize = 100000;
 pub struct P2PContext {
     pub tracker: TaskTracker,
     pub signer: SecretKey,
-    pub table: PeerTableHandle,
+    pub table: PeerTable,
     pub storage: Store,
     pub blockchain: Arc<Blockchain>,
-    pub(crate) broadcast: RLPxConnBroadcastSender,
+    pub(crate) broadcast: PeerConnBroadcastSender,
     pub local_node: Node,
     pub local_node_record: Arc<Mutex<NodeRecord>>,
     pub client_version: String,
@@ -55,7 +55,7 @@ impl P2PContext {
         local_node_record: Arc<Mutex<NodeRecord>>,
         tracker: TaskTracker,
         signer: SecretKey,
-        peer_table: PeerTableHandle,
+        peer_table: PeerTable,
         storage: Store,
         blockchain: Arc<Blockchain>,
         client_version: String,
@@ -145,7 +145,7 @@ pub(crate) async fn serve_p2p_requests(context: P2PContext) {
             continue;
         }
 
-        let _ = RLPxConnection::spawn_as_receiver(context.clone(), peer_addr, stream).await;
+        let _ = PeerConnection::spawn_as_receiver(context.clone(), peer_addr, stream).await;
     }
 }
 
@@ -158,17 +158,14 @@ fn listener(tcp_addr: SocketAddr) -> Result<TcpListener, io::Error> {
     tcp_socket.listen(50)
 }
 
-pub async fn periodically_show_peer_stats(
-    blockchain: Arc<Blockchain>,
-    mut peer_table: PeerTableHandle,
-) {
+pub async fn periodically_show_peer_stats(blockchain: Arc<Blockchain>, mut peer_table: PeerTable) {
     periodically_show_peer_stats_during_syncing(blockchain, &mut peer_table).await;
     periodically_show_peer_stats_after_sync(&mut peer_table).await;
 }
 
 pub async fn periodically_show_peer_stats_during_syncing(
     blockchain: Arc<Blockchain>,
-    peer_table: &mut PeerTableHandle,
+    peer_table: &mut PeerTable,
 ) {
     let start = std::time::Instant::now();
     loop {
@@ -186,7 +183,7 @@ pub async fn periodically_show_peer_stats_during_syncing(
             // Common metrics
             let elapsed = format_duration(start.elapsed());
             let peer_number = peer_table.peer_count().await.unwrap_or(0);
-            let current_step = METRICS.current_step.lock().await.clone();
+            let current_step = METRICS.current_step.get();
             let current_header_hash = *METRICS.sync_head_hash.lock().await;
 
             // Headers metrics
@@ -252,6 +249,7 @@ pub async fn periodically_show_peer_stats_during_syncing(
             // Storage leaves metrics
             let storage_leaves_downloaded =
                 METRICS.downloaded_storage_slots.load(Ordering::Relaxed);
+            let storage_accounts_inserted = METRICS.storage_tries_state_roots_computed.get();
             let storage_accounts = METRICS.storage_accounts_initial.load(Ordering::Relaxed);
             let storage_accounts_healed = METRICS.storage_accounts_healed.load(Ordering::Relaxed);
             let storage_leaves_time = format_duration({
@@ -356,7 +354,7 @@ headers progress: {headers_download_progress} (total: {headers_to_download}, dow
 account leaves download: {account_leaves_downloaded}, elapsed: {account_leaves_time}
 account leaves insertion: {account_leaves_inserted_percentage:.2}%, elapsed: {account_leaves_inserted_time}
 storage leaves download: {storage_leaves_downloaded}, elapsed: {storage_leaves_time}, initially accounts with storage {storage_accounts}, healed accounts {storage_accounts_healed} 
-storage leaves insertion: {storage_leaves_inserted_time}
+storage leaves insertion: {storage_accounts_inserted}, {storage_leaves_inserted_time}
 healing: global accounts healed {healed_accounts} global storage slots healed {healed_storages}, elapsed: {heal_time}, current throttle {heal_current_throttle}
 bytecodes progress: downloaded: {bytecodes_downloaded}, elapsed: {bytecodes_download_time})"
             );
@@ -366,7 +364,7 @@ bytecodes progress: downloaded: {bytecodes_downloaded}, elapsed: {bytecodes_down
 }
 
 /// Shows the amount of connected peers, active peers, and peers suitable for snap sync on a set interval
-pub async fn periodically_show_peer_stats_after_sync(peer_table: &mut PeerTableHandle) {
+pub async fn periodically_show_peer_stats_after_sync(peer_table: &mut PeerTable) {
     const INTERVAL_DURATION: tokio::time::Duration = tokio::time::Duration::from_secs(60);
     let mut interval = tokio::time::interval(INTERVAL_DURATION);
     loop {
@@ -374,12 +372,12 @@ pub async fn periodically_show_peer_stats_after_sync(peer_table: &mut PeerTableH
         let peers: Vec<PeerData> = peer_table.get_peers_data().await.unwrap_or(Vec::new());
         let active_peers = peers
             .iter()
-            .filter(|peer| -> bool { peer.channels.as_ref().is_some() })
+            .filter(|peer| -> bool { peer.connection.as_ref().is_some() })
             .count();
         let snap_active_peers = peers
             .iter()
             .filter(|peer| -> bool {
-                peer.channels.as_ref().is_some()
+                peer.connection.as_ref().is_some()
                     && SUPPORTED_SNAP_CAPABILITIES
                         .iter()
                         .any(|cap| peer.supported_capabilities.contains(cap))