+++ crates/networking/rpc/engine/payload.rs
@@ -604,7 +604,7 @@ async fn handle_new_payload_v1_v2(
     }
 
     // All checks passed, execute payload
-    let payload_status = try_execute_payload(&block, &context, latest_valid_hash).await?;
+    let payload_status = try_execute_payload(block, &context, latest_valid_hash).await?;
     Ok(payload_status)
 }
 
@@ -674,15 +674,17 @@ fn validate_block_hash(payload: &ExecutionPayload, block: &Block) -> Result<(), 
 }
 
 async fn try_execute_payload(
-    block: &Block,
+    block: Block,
     context: &RpcApiContext,
     latest_valid_hash: H256,
 ) -> Result<PayloadStatus, RpcErr> {
     let block_hash = block.hash();
     let block_number = block.header.number;
     let storage = &context.storage;
-    // Return the valid message directly if we have it.
-    if storage.get_block_by_hash(block_hash).await?.is_some() {
+    // Return the valid message directly if we already have it.
+    // We check for header only as we do not download the block bodies before the pivot during snap sync
+    // https://github.com/lambdaclass/ethrex/issues/1766
+    if storage.get_block_header_by_hash(block_hash)?.is_some() {
         return Ok(PayloadStatus::valid_with_hash(block_hash));
     }